exam.ai/requirements.txt
google-generativeai
streamlit
pypdf
python-dotenv
pandas
plotly